+++ packages/runtime/src/runtime.ts
@@ -9,6 +9,7 @@ import algosdk, {
 	Transaction,
 } from "algosdk";
 import { readFileSync } from "fs";
+import MD5 from "crypto-js/md5";
 import cloneDeep from "lodash.clonedeep";
 import nacl from "tweetnacl";
 
@@ -21,11 +22,12 @@ import { compareArray } from "./lib/compare";
 import {
 	ALGORAND_ACCOUNT_MIN_BALANCE,
 	ALGORAND_MAX_TX_ARRAY_LEN,
+	BlockFinalisationTime,
 	MAX_APP_PROGRAM_COST,
 	MaxExtraAppProgramPages,
+	seedLength,
 	TransactionTypeEnum,
-	ZERO_ADDRESS_STR
-} from "./lib/constants";
+	ZERO_ADDRESS_STR} from "./lib/constants";
 import { convertToString } from "./lib/parsing";
 import { LogicSigAccount } from "./logicsig";
 import { mockSuggestedParams } from "./mock/tx";
@@ -37,6 +39,7 @@ import {
 	ASADeploymentFlags,
 	ASAInfo,
 	AssetHoldingM,
+	Block,
 	Context,
 	EncTx,
 	ExecutionMode,
@@ -78,6 +81,7 @@ export class Runtime {
 			appCounter: ALGORAND_MAX_TX_ARRAY_LEN, // initialize app counter with 8
 			assetCounter: ALGORAND_MAX_TX_ARRAY_LEN, // initialize asset counter with 8
 			txReceipts: new Map<string, TxReceipt>(), // receipt of each transaction, i.e map of {txID: txReceipt}
+			blocks: new Map<number, Block>(),
 		};
 
 		this._defaultAccounts = this._setupDefaultAccounts();
@@ -90,7 +94,8 @@ export class Runtime {
 
 		// context for interpreter
 		this.ctx = new Ctx(cloneDeep(this.store), <EncTx>{}, [], [], this);
-		this.round = 2;
+		this.round = 2000;
+		this.populateChain(this.round);
 		this.timestamp = 1;
 	}
 
@@ -1184,7 +1189,7 @@ export class Runtime {
 	}
 
 	/**
-     * Parses the file and return the ABIContract in case of network.runtime not defined throw exception
+   * Parses the file and return the ABIContract in case of network.runtime not defined throw exception
 	 * @param fileName string
 	 * @retun parsed file
 	 */
@@ -1195,5 +1200,49 @@ export class Runtime {
 			throw new Error("ABI contract file is not valid");
 		}
 		return contract;
+
+	 * Produces new block and adds it to a Map where the keys are block numbers
+	 */
+	produceBlock() {
+		let timestamp: bigint | undefined;
+		let seed: Uint8Array | undefined;
+		if (this.store.blocks.size === 0) { //create genesis block
+			seed = nacl.randomBytes(seedLength);
+			timestamp = BigInt(Math.round(new Date().getTime() / 1000));
+		} else { //add another block
+			const lastBlock = this.store.blocks.get(this.round);
+			if (lastBlock) {
+				seed = new TextEncoder().encode(MD5(lastBlock.seed.toString()).toString());
+				timestamp = lastBlock.timestamp + BlockFinalisationTime;
+				this.round += 1;// we move to new a new round
+			}
+		}
+		if (timestamp && seed) {
+			this.store.blocks.set(this.round, { timestamp, seed });
+		} else {
+		throw new RuntimeError(RUNTIME_ERRORS.GENERAL.PRODUCE_BLOCK); 
+		}
+	}
+	/**
+	 * Returns a requested Block object. If it does not exist on chain throws an error
+	 * @param round block number
+	 * @returns Block
+	 */
+	getBlock(round: number): Block {
+		const block = cloneDeep(this.store.blocks.get(round));
+		if (block) {
+			return block;
+		}
+		throw new RuntimeError(RUNTIME_ERRORS.GENERAL.INVALID_BLOCK);
+	}
+	/**
+	 * Populates chain from first block to round number block (Produces N rounds) 
+	 * @param round current round number
+	 */
+	private populateChain(round:number) {
+		this.round = 1
+		for(let blockN = 1; blockN<=round; blockN++){
+			this.produceBlock();
+		}
 	}
 }